chore: move the definition of `allocate_pages` into `allocate` module
To use this method for implementing syscall.
